pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-cors>=6.0.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.86.0",
    "psycopg2-binary>=2.9.10",
    "psutil>=7.0.0",
    "schedule>=1.2.2",
    "requests>=2.32.4",
    "cryptography>=45.0.4",
    "pyjwt>=2.10.1",
]
